Cleaned up invites models
