+++ models/Invites/Invites.ts
@@ -1,0 +1,8 @@
+import create from "./create";
+import remove from "./delete";
+import get from "./get";
+import join from "./join";
+export const createInvite = create;
+export const deleteInvite = remove;
+export const getInviteById = get;
+export const joinOrgFromInvite = join;
