+++ models/Invites/join.ts
@@ -1,0 +1,69 @@
+import {
+  TransactWriteCommandInput,
+  TransactWriteCommand,
+} from "@aws-sdk/lib-dynamodb";
+import { Dynamo } from "../../awsClients/ddbDocClient";
+import { ENTITY_TYPES } from "../../Config";
+import { JoinOrgFromInviteInput } from "../../types/main";
+import * as Time from "../../utils/time";
+const { DYNAMO_TABLE_NAME } = process.env;
+export default async function join(
+  props: JoinOrgFromInviteInput
+): Promise<void> {
+  const { userId, invite } = props;
+  // TODO types
+  try {
+    const transactParams: TransactWriteCommandInput = {
+      TransactItems: [
+        {
+          // Delete org invite
+          Delete: {
+            Key: {
+              PK: `${ENTITY_TYPES.USER}#${userId}`,
+              SK: `${ENTITY_TYPES.ORG_INVITE}#${invite.inviteId}`,
+            },
+            TableName: DYNAMO_TABLE_NAME,
+          },
+        },
+
+        {
+          // Update the user with the new org
+          Update: {
+            Key: {
+              PK: `${ENTITY_TYPES.USER}#${userId}`,
+              SK: ENTITY_TYPES.USER,
+            },
+            TableName: DYNAMO_TABLE_NAME,
+            UpdateExpression:
+              "SET orgId = :orgId, orgJoinDate = :orgJoinDate, GSI1PK = :GSI1PK",
+            ExpressionAttributeValues: {
+              ":orgId": invite.orgId,
+              ":orgJoinDate": Time.currentISO(),
+              ":GSI1PK": `${ENTITY_TYPES.ORG}#${invite.orgId}#${ENTITY_TYPES.USER}S`,
+            },
+          },
+        },
+        {
+          // Increment the org with the new user
+          Update: {
+            Key: {
+              PK: `${ENTITY_TYPES.ORG}#${invite.orgId}`,
+              SK: ENTITY_TYPES.ORG,
+            },
+            TableName: DYNAMO_TABLE_NAME,
+            UpdateExpression: "SET totalUsers = totalUsers + :value",
+            ExpressionAttributeValues: {
+              ":value": 1,
+            },
+          },
+        },
+      ],
+    };
+
+    await Dynamo.send(new TransactWriteCommand(transactParams));
+    return;
+  } catch (error) {
+    console.error(error);
+    throw new Error(error);
+  }
+}
